Conflicts:
	java/code/src/com/redhat/rhn/common/db/datasource/xml/Errata_queries.xml
	java/code/src/com/redhat/rhn/domain/channel/Channel.hbm.xml
	spacewalk/admin/rhn-populate-database.pl
	spacewalk/setup/bin/spacewalk-setup

+++ java/code/src/com/redhat/rhn/manager/system/UpdateBaseChannelCommand.java
@@ -76,15 +76,14 @@ public class UpdateBaseChannelCommand extends BaseUpdateChannelCommand {
                         server, newChannel)) {
             return new ValidatorError("system.channel.nochannelslots");
         }
-        List <Integer> newKidsToSubscribe = new LinkedList<Integer>();
+        List <Long> newKidsToSubscribe = new LinkedList<Long>();
         
         if (oldChannel != null && newChannel != null) {
             Map<Channel, Channel> preservableChildren = ChannelManager.
                             findCompatibleChildren(oldChannel, newChannel, user);
             for (Channel kid : server.getChannels()) {
                 if (preservableChildren.containsKey(kid)) {
-                    newKidsToSubscribe.add(preservableChildren.get(kid).
-                            getId().intValue());    
+                    newKidsToSubscribe.add(preservableChildren.get(kid).getId());
                 }
             }
         }
